Refactor and modernize test code for clarity and consistency
Refactored test suite for improved readability, maintainability, and compliance:
- Replaced positional with named arguments in method calls and assertions.
- Updated reflection, buffer, and stream operations for explicitness.
- Unified exception handling and assertion patterns.
- Improved async test cancellation, resource management, and coverage.
- Added/updated region markers, comments, and interface implementation notes.
- Cleaned up formatting, removed redundant/useless code and usings.
- Ensured proper disposal, awaited async code, and handled exceptions.
- Updated copyright/license headers.
No changes to core library logic; all updates are test-only.

+++ test/Xtate.Core.Test/UnitTests/Common/StackSpanAndSafeFactoryCoverageTest.cs
@@ -1,17 +1,17 @@
 // Copyright © 2019-2026 Sergii Artemenko
-//
+// 
 // This file is part of the Xtate project. <https://xtate.net/>
-//
+// 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as published
 // by the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
-//
+// 
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU Affero General Public License for more details.
-//
+// 
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
@@ -29,16 +29,16 @@ public class StackSpanAndSafeFactoryCoverageTest
 
 		Assert.IsTrue(!small);
 		Assert.IsFalse(small ? true : false);
-		Assert.AreEqual(StackSpan<int>.MaxLengthInStack, (int) small);
+		Assert.AreEqual(StackSpan<int>.MaxLengthInStack, small);
 		small.Dispose();
-		Assert.ThrowsExactly<InvalidOperationException>([ExcludeFromCodeCoverage] () => ConvertSmallStackSpanToSpan());
+		Assert.ThrowsExactly<InvalidOperationException>([ExcludeFromCodeCoverage]() => ConvertSmallStackSpanToSpan());
 
 		var requestedLength = StackSpan<int>.MaxLengthInStack + 1;
 		var large = new StackSpan<int>(requestedLength);
 
 		Assert.IsFalse(!large);
 		Assert.IsTrue(large ? true : false);
-		Assert.AreEqual(requestedLength, (int) large);
+		Assert.AreEqual(requestedLength, large);
 		Span<int> span = large;
 		Assert.AreEqual(requestedLength, span.Length);
 		span[0] = 17;
